requirements.txt
# Core
fastapi>=0.115.0
uvicorn>=0.30.6
pydantic>=2.9.2
pyyaml>=6.0.2
python-dotenv>=1.0.1

# AI – Chat (Groq free tier, Llama 3.3 70B)
groq>=0.11.0

# AI – Embeddings (local, no API key needed)
sentence-transformers>=3.0.0

# Vector Search
faiss-cpu>=1.12.0
numpy>=1.26.4

# Knowledge Graph
networkx>=3.3

# HTTP
httpx>=0.27.2

# HTML parsing (Hyland Docs connector)
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Logging
rich>=13.8.1

# CLI
typer>=0.12.5
